chore: Add more fine grained metrics for the http outcalls price  (#5609)
This is an extension of https://github.com/dfinity/ic/pull/5562.

Instead of recording metrics with specific constants, we log all
parameters needed for the cost formula in order to be able to quickly
adjust the weights in the formula during analysis, without needing a new
release each time.

There are about 60 http requests per seconds across all subnets and
nodes; so this would add about 36 million documents to our elastic
search, which should be manageable (given they are only stored for a
week).