feat(sns): SNS treasury manager API (#5788)
This PR adds the Rust API types for the **SNS Treasury Manager**
extensions.

# Why

For the [upcoming SNS Liquidity Pools
feature](https://forum.dfinity.org/t/sns-liquidity-pools-design/50439),
the design involves implementing SNS extensions for managing assets on
behalf of the DAO; concretely, by allocating some of the treasury tokens
into DEX liquidity pools.

The first such extension is planned to be implemented as a separate
canister, as it needs to have DEX-specific code that should not be part
of the ICP mono repo. This is why the SNS and the SNS extension
implementation need an API to agree on. This is that API, although we
make it more general and not tailor specifically to managing DEX assets,
so that it suites potential future use cases.

# What

The main addition is the Candid file which should be seen as the source
of truth for the API.

In addition, there are Rust types to aid implementors and a mock
canister implementation that is used to test that those Rust types
exactly match the Candid service definitions.

---------

Co-authored-by: ShahriarJavidi <shahriar.javidi@dfinity.org>
Co-authored-by: ShahriarJavidi <33692111+ShahriarJavidi@users.noreply.github.com>