test(bitcoin): Adapter system tests (#5575)
This PR adds bitcoin system tests.

In particular, this PR ports 3 existing adapter unit tests (found in
`rs/bitcoin/adapter/tests/adapter_tests.rs`) into system tests. These
are:

- `test_received_blocks`
- `test_receives_new_3rd_party_txs`
- `test_send_tx`


### Test setup

The test setup consists of a `AdapterProxy` structure which spawns a
`MessageCanister` at the `PrincipalId` of the bitcoin canister, which
proxies the protected `bitcoin_send_tx` and `bitcoin_get_successors`
calls from the test to the replica. The effect is that we can test
arbitrary logic from the test environment, which exercises the entire
code path from the canister through execution, consensus and the adapter
client to the adapter itself, while the adapter unit tests only tests
client and adapter.

The test reuses parts of the setup of the ckbtc tests, in particular,
spawning a `bitcoind` inside a `UnivervsalVM`, to act as the
counter-party for the adapter to talk to. In the unit test, bitcoind was
simply executed as a binary.

---------

Co-authored-by: gregorydemay <112856886+gregorydemay@users.noreply.github.com>