fix(test): fix the signer canister and revert chain key performance tests back to it  (#5853)
The introduction of the signer canister (#5530) was quickly reverted
(#5786) because

> responses [...] are currently always interpreted as successful, even
if the actual signature request failed

The message canister and signer canister interfaces are written
identically and it is surprising to see that they behave differently. In
reality though, we use a pre-built wasm file for the message canister
which [has not been
modified](https://github.com/dfinity/ic/commits/master/rs/tests/src/message.wasm?browsing_rename_history=true&new_path=rs/tests/driver/src/message.wasm&original_branch=b509aaa1ffb891603a19b5133dd8b7b57f0a6444)
since it was created. Though, this is [not the
case](https://github.com/dfinity/ic/commits/master/rs/tests/test_canisters/message/src/main.rs)
for the Rust source code of the message canister. So the canister which
is used in the tests is actually one corresponding to the original
source code and the current Rust code is not in sync with the built
wasm. In particular, PR #4024 changed the canister's interface to not
propagate the reject of the inter-canister call back to the user, but
this change is actually not reflected in the tests, since we haven't
rebuilt the wasm.

This PR thus changes the interface of the signer canister to mirror the
one that the message canister has inside the built wasm, i.e. the one
without the changes from #4024. It thus correctly rejects the call in
case the inter-canister call itself got rejected.

An example of a test run with this new signer canister can be found
[here](https://dash.zh1-idx1.dfinity.network/invocation/1bef5f83-9ff0-449d-96fc-a7ee4792bf3e#@4866).
You can now see that some requests are successfully failing with about
the same success rate (~40-45%) as with a [run from current
master](https://dash.zh1-idx1.dfinity.network/invocation/68cb8102-ad08-44df-8c41-537a11c2d745#@4444),
which uses the message canister.


FYI: I am also working on #5861 to get rid of the wasm file and to build
the message canister at every test, just like the signer canister, in
order to match with the source code.