chore: simplify btc adapter build graph
This simplifies the btc adapter build graph to speed up builds and
to have fewer rebuilds.

In particular, the dependency list is made more explicit, and unused
dependencies are removed.

Moreover, binaries (both standalone and test binaries) are moved out of
the crate's glob-ed sources to make sure they do not need to recompile
every module.