fix(test): fix Rust SignerCanister API (#5891)
#5853 just got merged and led to a
[regression](https://github.com/dfinity/ic/actions/runs/16197096606/job/45726302061).
It fixed the signer canister interface to always return a signature, or
reject the call if that is not possible, as opposed to returning an
`Ok`/`Err` but I had forgotten to adapt the Rust `SignerCanister` API to
reflect this change. I'll use `CI_ALL_BAZEL_TARGETS` next time...

This PR fixes this. Now the API of `MessageCanister` and
`SignerCanister` are similar, since `SignerCanister` no longer returns a
`Result` (i.e. `forward_to` and
`gen_ecdsa_sig`/`gen_schnorr_sig`/`gen_vetkd_key` have the same function
signatures).