feat(nns-governance): Introduce fulfill subnet rental request proposal. (#5835)
# Background

This is part of the "subnet rental" feature, approved in [motion
proposal 128820][approved].

[approved]: https://dashboard.internetcomputer.org/proposal/128820

# Overview

This new proposal type is behind a feature flag. More precisely, if
someone tries to make such a proposal, they will get an
`InvalidProposal` error. Therefore, there are no "real" behavior changes
here (in release builds), but there will be, once enabled. OTOH, with
this change, the Governance canister does know how to deserialize such
proposals (even in release builds); they will just get rejected until it
is enabled.

One reason that we gate this behind a flag is to let clients (esp NNS
Dapp and IC Dashboard) prepare for the existence of such proposals. In
the meantime, even though this new type of proposal is in the Candid
interface of the Governance canister, clients need not worry that they
will see such proposals in responses (until they are enabled).

## Proposal Behavior

(This section assumes that the new proposal type has been enabled.)

The main thing that this new proposal type does is create a "rented"
subnet. Such creation is based on a prior "rental request" in the
[Subnet Rental canister]. Some special things about this subnet:

1. Only the "user" (from the rental request) can create canisters in the
subnet.

2. Canisters are not charged (cycles) by the platform.

[Subnet Rental canister]:
https://github.com/dfinity/subnet-rental-canister

The other thing that this new proposal type does is notify the Subnet
Rental canister that the subnet has been created in order to fulfill the
existing rental request. This has a couple of effects:

1. The rental request becomes a "rental agreement".

2. The user is authorized to create canisters in the subnet. (To
implement this, the Subnet Rental canister calls the Cycles Minting
canister.)

# Future Work

This change does NOT include an implementation of the new proposal type;
but it does include validation. Implementation will be done later. Not
doing it here is fine, because making such proposals is blocked anyway.

Since the validation code is not reachable, there are no tests for that
in this change. Instead, tests will be added later.

# References

Begins to address [NNS1-3925], but does not close.

[NNS1-3925]: https://dfinity.atlassian.net/browse/NNS1-3925

[👈 Previous PR][prev] | [Next PR 👉][next]

[prev]: https://github.com/dfinity/ic/pull/5816
[next]: https://github.com/dfinity/ic/pull/5885