chore(CRP-2613): Remove deprecated compute_initial_i_dkg_dealings endpoint Pt.1 (#5866)
This PR deactivates the `compute_initial_i_dkg_dealings` endpoint from
the management canister, since it has been superseded by the
`reshare_chain_key` endpoint, and the registry canister has been
migrated to use the new endpoint.

In particular this PR:
- Makes all calls to `compute_initial_i_dkg_dealings` fail
- Migrates some leftover test-cases to use `reshare_chain_key` calls
instead
- Removes the routing logic of the call, and instead directs the call to
the own subnet

In a followup PR, I will remove the logic from consensus to process the
call. Since this is an internal API and we know that it has not been in
use since the registry canister was migrated, we know that there are no
open contexts and the whole removal could happen in one PR, however, I
still think it makes sense to split this into 2 PRs, to be easier to
review.