feat[MR-670]: New format for canister ranges in state tree
This PR adds the /canister_ranges subtree into the state tree. This
subtree encodes the routing table in a sharded format where each leaf only
contains a small number of canister ranges.

The existing, unsharded way to encode routing tables under
/subnet/<canister_id>/canister_ranges is still supported by the state
tree and will remain available for backwards compatibility. In a
follow-up the new format will in particular be used for NNS delegations.