chore: remove max canister memory size if no memory allocation is set (#5775)
This PR removes the notion of canister memory limit (a limit on the
total canister memory usage) if no explicit memory allocation is set.
The value used to be the sum of the maximum wasm and stable memory sizes
which is not a meaningful bound on the total canister memory usage which
also includes (potentially multiple) canister snapshots and other pieces
of canister state. Note that wasm and stable memory sizes are still
subject to their respective limits (4GiB for wasm32, 6GiB for wasm64,
and 500GiB for stable), this PR only drops a check on the total memory
usage of a canister with best-effort memory allocation which was used in
invariant checks to raise alerts (i.e., not directly affecting the
protocol behavior).

In more detail, this PR performs the following cleanup:
- removes `canister_memory_limit` from `ExecutionParameters`;
- removes `max_canister_memory_size_wasm32` and
`max_canister_memory_size_wasm64` from `execution_environment::Config`;
- removes `limit` from `MemoryUsage`;
- consolidates memory usage checks in canister state invariant checks;
- drops obsolete tests and hardens existing tests.