feat: CON-1528 Signature request context can hold the full pre-signature (#5862)
https://github.com/dfinity/ic/pull/5768 created a collection in
replicated state that will be used to store available tECDSA and
tSchnorr pre-signatures.

These pre-signatures are consumed by incoming signature requests.
Specifically, signature request contexts are matched with available
pre-signatures, by moving one pre-signature out of the pre-signature
stash and including it in the context.

This PR extends the signature requests contexts for ECDSA and Schnorr to
optionally hold this matched pre-signature. When the context is created
initially, the pre-signature is set to `None`. The actual pairing (once
a pre-signature is available) is done in a later step.

To ensure version compatibility, this field will continue to be set to
`None` for now, until this change was rolled out to all subnets.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>