fix: mount back ict_testnets dir into the local container (#5942)
Previously we used to mount the whole `/tmp` dir into the container
image which was pretty convenient for sharing files among containers.
after #5601 we stopped doing that because it conflicts with other
requirements.

This PR mounts back at least `/tmp/ict_testnets` which is used as
default dir for ict log output.