test(CON-1442): Use well formed keys to test IC00Methods (#5906)
This PR adds well formed chain keys to the unit test setup for
`/rs/execution_environment`.

As mentioned in #2693, before we used fixed random bytes (such as
`abababab`) as the keys. This means, that the `ic00` calls to
`VetKDPublicKey`, `SchnorrPublicKey` and `ECDSAPublicKey` could not
succeed, as the crypto side would fail to parse the keys.

This PR replaces the fixed random bytes with fixed random keys.
It also adds two missing enpoint tests.

This increases test coverage, as the happy paths of these endpoints are
covered by unit tests now.