chore: switch from boot partition version to /opt/ic/share/version.txt (#5935)
With the recent issue, we observed that the version.txt on the boot
partition is not a good indicator of a successful upgrade. The boot
partition is mounted later, where as root (with /opt/ic/share) is
mounted very early on. In that issue, the system failed to "switch over"
to the alternate root, but would then mount the right boot.