feat: adding vector vm for scraping logs from testnets (#5948)
This PR adds a small vector vm that can be configured to fetch logs from
a testnet. The idea is to bridge the gap to production setup where we
pull logs from the nodes and push them to elastic. In following PR's we
would be able to remove `filebeat` and make the images slimmer.