feat: Build custom SEV-compatible OVMF from source and distribute it with the GuestOS (roll forward) (#5947)
The custom OVMF is necessary to support SEV-SNP. Since the OVMF is used
for the calculation of the SEV measurement, we include it in the Guest.
This has the benefit that we get a consistent measurement for each
GuestOS release and we don't have to keep track of all OVMFs that may be
lying around in HostOS-es.

This PR was already submitted previously as #5776 which had to be rolled
back because it broke direct boot. It turns out that the SEV-specific
OVMF should only be used when booting with SEV. So the amended logic
uses the regular OVMF from the HostOS when we boot without SEV (or the
GuestOS does not contain the OVMF).

The first "Revert" commit is just the original code from #5776. The rest
of the commits are new.

---------

Co-authored-by: IDX GitHub Automation <>