feat: CON-1531 Retain `IDkgTranscript`s stored in replicated state (#5925)
Before an `IDkgTranscript` may be used, i.e. to create a signature
share, it needs to be loaded into the crypto component via the
`IDkgProtocol::load_transcript` API. Conversely, transcripts are purged
from the crypto component periodically via the
`IDkgProtocol::retain_active_transcripts` API.

In order to successfully answer all signature requests, it is therefore
important that no transcripts are purged too early. Currently, each IDKG
transcript is stored on the blockchain until it is consumed by a
signature requests. Therefore, retaining only these transcripts on the
blockchain is enough to satisfy the requirement.

In the future however, we want to start storing pre-signatures and their
corresponding transcripts in the replicated state. At that point, there
may be signature requests that were paired to pre-signatures and key
transcripts which exclusively exist in the replicated state.

This PR therefore makes sure to additionally retain any transcripts that
were matched to ongoing signature requests in certified state.

A special case are key rotations: During a key rotation there may be a
point in time where the old key transcript no longer exists on the
blockchain, but it was still paired to an ongoing signature request that
hasn't been certified yet. Therefore, we additionally retain the key
transcripts of all pre-signature stashes in the latest certified state,
in case these key transcripts were paired with signature requests that
haven't been certified yet.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>