feat(nns-governance): Execute FulfillSubnetRentalRequest proposals. (#5854)
In an [earlier PR], we introduced this new proposal type, but did not
implement (the execution of) it. Instead, execution is implemented here.

[earlier PR]: https://github.com/dfinity/ic/pull/5835

# Future Work

This is not tested yet. This is ok, because the code is behind a flag.
Rest assured that end to end test(s) will be written before the flag is
set to "enabled".

Need to confirm that this correctly sets the fields of
`CreateSubnetPayload`. The values in this PR were taken from "the wild",
i.e. an executed proposal.

Want to source Subnet Rental canister request and response types from an
official library. I don't think that exists yet. It could come from the
[subnet-rental-canister repo][src], or we could have our own here in the
ic repo.

[src]: https://github.com/dfinity/subnet-rental-canister

# References

Closes [NNS1-3925].

[NNS1-3925]: https://dfinity.atlassian.net/browse/NNS1-3925

[👈 Previous PR][prev]

[prev]: https://github.com/dfinity/ic/pull/5895