feat(nns): Keep the voting power spike detection when the snapshots are not full (#5932)
# Why

If the voting power spike detection mechanism has a false positive, we
will need to remove most of the voting power snapshots so that new
voting power will be considered normal. In that case, we don't want the
detection mechanism to be inactive, which would be the case if we only
detect spikes if the voting power snapshots are full. The "fullness"
check was added to get around the check in tests, but as
@andrew-lee-work pointed out, we can limit the check for tests.

# What

In production code, do not skip the voting power spike detection when
the snapshots are not full.