fix: fail when IcGatewayVm does not become healthy (#5988)
Bug fix
===
Previously when the IcGatewayVm did not become healthy it only logged an
error message but did not in fact fail. This caused tests to fail later
in unexpected ways causing hard to debug flakiness (for example in our
most flaky test `//rs/tests/sdk:dfx_smoke_test` which flakes 8.5% of the
time).

So now the call to `IcGatewayVm::new(...).start(...)` will fail when the
IcGatewayVm does not become healthy in time.


DNS propagation
===

During testing we also observed, multiple times, that the DNS records
were not yet propagated after registering them with Route53:

```
2025-07-21 12:09:25.942 INFO[setup:rs/tests/driver/src/driver/ic_gateway_vm.rs:410:0] Waiting for IcGatewayVm to become healthy ...
...
2025-07-21 12:14:21.833 DEBG[setup:rs/tests/driver/src/driver/test_env_api.rs:2210:0] Func="await_status_is_healthy of ic-gateway with url https://ic2.farm.dfinity.systems/api/v2/status [rs/tests/driver/src/driver/ic_gateway_vm.rs:412]" failed on attempt 60. Error: failed to execute request\n \n Caused by:\n     0: error sending request for url (https://ic2.farm.dfinity.systems/api/v2/status)\n     1: client error (Connect)\n     2: dns error: no record found fo...
2025-07-21 12:14:26.912 DEBG[setup:rs/tests/driver/src/driver/test_env_api.rs:2195:0] Func="await_status_is_healthy of ic-gateway with url https://ic2.farm.dfinity.systems/api/v2/status [rs/tests/driver/src/driver/ic_gateway_vm.rs:412]" succeeded after 300.970602165s on attempt 61
```

Since it took 300 seconds for the IcGatewayVm to become healthy the
hypothesis was that this was caused by some recursive resolver caching a
negative lookup for 300 seconds after which the record resolved
correctly.

To wait for complete DNS propagation this commit also repeatedly checks
a random domain with the registered playnet suffix (for example:
`iJwzMa2y.ic0.farm.dfinity.systems`). This should work because we
register a wildcard for the playnet.

However, the configured TTL in and of the [SOA
record](https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/SOA-NSrecords.html)
record is actually 900 seconds. So I'm unsure where the 300 seconds is
coming from. In any case, the wait for DNS propagation does seem to work
([also on
CI](https://dash.zh1-idx1.dfinity.network/invocation/40f52088-462e-43ac-82fc-288bca027420?target=%2F%2Frs%2Ftests%2Fsdk%3Adfx_smoke_test&targetStatus=5#@1399)):

```
2025-07-21 19:19:05.813 INFO[setup:rs/tests/driver/src/driver/ic_gateway_vm.rs:422:0] Waiting for DNS propagation of wildcard records for domain: ic0.farm.dfinity.systems
2025-07-21 19:19:05.813 DEBG[setup:rs/tests/driver/src/driver/test_env_api.rs:2188:0] Func="DNS propagation check for domain ic0.farm.dfinity.systems [rs/tests/driver/src/driver/ic_gateway_vm.rs:428]" is being retried for the maximum of 600s with a constant backoff of 5s
2025-07-21 19:19:05.826 DEBG[setup:rs/tests/driver/src/driver/test_env_api.rs:2210:0] Func="DNS propagation check for domain ic0.farm.dfinity.systems [rs/tests/driver/src/driver/ic_gateway_vm.rs:428]" failed on attempt 1. Error: DNS lookup failed for iJwzMa2y.ic0.farm.dfinity.systems: failed to lookup address information: No address associated with hostname\n \n Stack backtrace:\n    0: anyhow::error::<impl anyhow::Error>::ms...
2025-07-21 19:19:10.846 INFO[setup:rs/tests/driver/src/driver/ic_gateway_vm.rs:441:0] DNS propagation confirmed: fhP8XQV0.ic0.farm.dfinity.systems resolves correctly
2025-07-21 19:19:10.846 DEBG[setup:rs/tests/driver/src/driver/test_env_api.rs:2195:0] Func="DNS propagation check for domain ic0.farm.dfinity.systems [rs/tests/driver/src/driver/ic_gateway_vm.rs:428]" succeeded after 5.033702409s on attempt 2
```

API Boundary Node Health Check
===

Since the IcGatewayVm connects to the API Boundary Nodes it's important
the latter are healthy before we start the former. So this commit also
adds a `node.await_status_is_healthy()` for all API boundary nodes
before the `start_gateway_container(...)`.


Future work
===

Even with these improvements. I have seen it take quite some time before
the IcGatewayVm becomes healthy (the longest I've seen was 5 minutes):

```
2025-07-21 19:01:20.473 INFO[setup:rs/tests/driver/src/driver/ic_gateway_vm.rs:117:0] Waiting for all API boundary nodes to become healthy ...
...
2025-07-21 19:03:16.810 DEBG[setup:rs/tests/driver/src/driver/test_env_api.rs:2210:0] Func="await_status_is_healthy of ic-gateway with url https://ic0.farm.dfinity.systems/api/v2/status [rs/tests/driver/src/driver/ic_gateway_vm.rs:462]" failed on attempt 13. Error: failed to execute request\n \n Caused by:\n     0: error sending request for url (https://ic0.farm.dfinity.systems/api/v2/status)\n     1: client error (Connect)\n     2: tcp connect error: No route t...
...
2025-07-21 19:03:21.892 DEBG[setup:rs/tests/driver/src/driver/test_env_api.rs:2195:0] Func="await_status_is_healthy of ic-gateway with url https://ic0.farm.dfinity.systems/api/v2/status [rs/tests/driver/src/driver/ic_gateway_vm.rs:462]" succeeded after 119.977433515s on attempt 14
``` 

For this it would be good to have extra logs of the ic-gateway to see if
it causes the "no route to host" error for example due to not listening
or accepting connections yet.