feat: add bazel command to diff script (#6014)
This makes the `diff.sh` script (used to shortcut builds) aware of the
bazel command being executed. This means that it will only return `test`
targets if the command is `test`, as `bazel test <targets>` will fail if
called with exclusively `build` targets.

PRs might still skip the build if no _test_ needs to be rebuilt, even
though some build targets should be rebuilt. This is however not a new
issue.