test(nns): Refactor golden state test to more easily add more sanity checks (#6006)
To make it easier to add new sanity checks, refactor the existing sanity
check to be more reusable.


Some changes in addition to refactoring:

* Compare the "after" against "before upgrades", rather than "before
advancing time". This should be considered more defensive, since
anomalies can happen right after the upgrade
* Perform sanity checks whether NNS Governance is upgraded or not. This
is simpler, and does not affect CI runs (as CI runs always upgrade
governance). It also makes it easier to check for other canisters in the
future.