chore: Add more Http outcalls logs (#5985)
**This is just a temporary change, and all logs will be removed after a
couple of weeks**
We need the following parameters for every http request:
- request size
- downloaded response size
- response time
- `max_response_bytes` parameter
- block payload size
- subnet size

All of those parameters except for the payload size are known before
consensus, however, the block payload size is only known after the
relevant block is finalized. Therefore there are two ways of "coupling"
the logs:
1. Store all of the mentioned parameters in some cache memory, and log
all of them after consensus. **This adds a lot of complexity to the
code**
2. Have two log statements, and couple them by request ID, process ID
(because on each replica restart, the request ID counter is also reset)
and perhaps log timestamp (they have to be within 1-2 minutes).

In the end I decided to go with approach two as it's more pragmatic.
This implies that we will double the total documents we have stored for
http outcalls.
[Currently](https://kibana.mainnet.dfinity.network/app/r/s/jkMi5) we
have around 5 million new logs every day. **This change would push that
to around 10 million.**