feat(ICRC_Rosetta): FI-1794: Handle startup ledger errors (#5884)
Do not fail multi-token ICRC Rosetta startup as long as setup of at
least one token is successful.

With the introduction of support for multiple ICRC tokens in a single
Rosetta instance, Rosetta currently fails during startup in case the
setup of any of the configured tokens fails. In particular, this can
happen if the token metadata cannot be retrieved from the ledger, e.g.,
if the canister has run out of cycles. This PR adds more leniency during
the startup process - if the setup of any of the configured tokens
fails, a warning is logged. In case the setup of all configured tokens
fails, Rosetta exist with an error, but if the setup of at least one
token succeeds, the Rosetta instance continues running. It is then up to
the admin to examine the logs and identify issues.