fix: Check input queue before canister migration
Before performing a canister migration, we check that the canister is
ready for migration, i.e. it is stopped and has flushed all of the relevant messages. This PR fixes a
bug by also checking that the input queue is empty, as it is possible
for a canister to be both stopped and have a non-empty input queue.

This bug had no effect on production as canister migrations are not
enabled yet.