feat(idkg): [CON-1529] add `max_pre_signature_transcripts_in_creation` to `ChainKeyConfig` (#5972)
This PR introduces the `max_pre_signature_transcripts_in_creation`
parameter to the `ChainKeyConfig` of a subnet. It controls how many
pre-signature transcripts a subnet can work on in parallel.

Currently, we use `KeyConfig.max_pre_signature_transcripts_in_creation`
for that, while also using it for the size of the stash. This means that
all pre-signatures are regenerated at once if the stash is empty. As
part of the ongoing improvements to threshold signatures, pre-signatures
will be stored in the replicated state instead of blocks, which will
allow us to increase the size of the stash.

However, it would be infeasible to create all of them at once, so this
new parameter will let us control how many are created in parallel.

For compatibility, the parameter is added as optional and defaults to
`None`.