chore: support retrieving the hostname of IC nodes in testnets (#6029)
To implement [this
suggestion](https://github.com/dfinity/ic/pull/6011#discussion_r2222826085),
besides the existing:
`InternetComputer::setup_and_start(&mut self, env: &TestEnv) ->
Result<()>` this commit introduces:
```rust
InternetComputer::setup_and_start_return_vms(
    &mut self,
    env: &TestEnv,
) -> Result<BTreeMap<String, AllocatedVm>>
```
Which returns information on the allocated VM per IC node (keyed on the
name of the VM which corresponds to the Node ID).

The `AllocatedVm` is extended with a `hostname` field that denotes on
which host the VM was allocated.