test: [MR-697] testnet can add nodes with different specified hosts (#6005)
I've been testing super large state size in a testnet with multiple
nodes, where each node is deployed to a dedicated performance host to
match production-level I/O performance. This PR introduces the ability
to specify required host features for each node, enabling customized,
per-node deployment.

It also includes a system test that creates such a testnet. The test
reads hostnames from environment variables and leverages the newly added
functionality to streamline the performance tests we have been running.

Note: Migrating the image to the physical devices is still a manual
step. Automation for this process will be introduced in a separate PR.