test(recovery): guestos-recovery-engine smoke test (#5761)
This PR tests the guestos-recovery-engine component of NNS recovery. It
introduces a new CI job ([example
run](https://github.com/dfinity/ic/actions/runs/15969553403/job/45037480353))
that:
- Creates a dummy CUP (random content)
- Creates a dummy registry local store (random content), but still
preserves some kind of expected structure
- Creates an archive `recovery.tar.zst` of those
- Uploads it to the upstreams
- Minor detail: As the URL is always the same during this test (version
`000...000`), we explicitly tell CDNs not to cache it as we want to be
sure we have the newest archive (see the changes in
`ci/src/artifacts/upload.sh`).
- Sets the `VERSION` and `EXPECTED_RECOVERY_HASH` variables in
`guestos-recovery-engine.sh`
- Runs the new test
`//rs/tests/consensus:guestos_recovery_engine_smoke_test`

This new test creates an IC with a single subnet of a single node that
uses the recovery image. This image will execute
`guestos-recovery-engine.sh`, thus fetching the artifacts from the
upstreams and setting its CUP and registry local store. The test
directly opens a SSH connection to the node, even before it is healthy
(it will in fact never be healthy as the orchestrator will always panic
before displaying its dashboard because it will not find a configured
NNS public key in the local store, as the content of the latter is
garbage). It checks the content of the CUP and of the local store and
succeeds if they match the original ones, passed as environment
variables to the test.

As the content of these files are generated completely randomly, it will
almost never be valid UTF-8. Also, it could contain special characters
like \x00, \x0a or \x20 which could go wrong in a lot of places (when
storing them in envvars, inside the CI job because it will try to print
the envvars in question, in the SSH command of the test because it
expects UTF-8). For these reasons, we store the base64 encoding of the
files in the variables and compare these, instead of the raw binary
files.

@dfinity/idx I am happy to take your critical feedback on the changes in
`ci/src/artifacts/upload.sh` and inside the new CI job
(`.github/workflows-source/schedule-daily.yml`), whether I am not
forgetting anything and/or am following good practices, as this is my
first time playing with GitHub Actions.

---------

Co-authored-by: Andrew Battat <andrew.battat@dfinity.org>
Co-authored-by: IDX GitHub Automation <>
Co-authored-by: Nicolas Mattia <nicolas.mattia@dfinity.org>