feat(governance): Add validation and fix display for reward_account (#6033)
This adds validation for node provider reward accounts to ensure they
include checksums (i.e. are more likely to be valid accounts) and fixes
the outputs to always display the 32 bit version with checksums, which
match the normal account identifier display are more readily searched
for in the dashboard and with other tools and APIs.