feat: CON-1526 Allow full pre-signatures to be delivered with the batch (#6019)
Currently, consensus only includes the _IDs_ of available pre-signatures
in the batch delivered to execution, where they are paired with incoming
signature request contexts.

This PR changes the type to instead deliver a map `Map<PreSigId,
Option<PreSignature>>`, in preparation for delivering full
pre-signatures to execution. For now, all values inserted into this map
are still `None`. Therefore, the logic is unchanged compared to before.

Additionally, we deliver all current IDKG key transcripts. This will
become relevant once we start maintaining a stash of pre-signatures in
replicated state, because the stash needs to be purged whenever the key
transcript changes.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>