feat: [EXC-2080] Make cycles charging dependent on registry flag  (#5847)
### Why
In [rental
subnets](https://forum.dfinity.org/t/subnet-rental-canister/28334/9),
canisters are not charged cycles for canister creation, execution,
memory etc.
The registry flag `canister_cycles_cost_schedule` (introduced in
https://github.com/dfinity/ic/pull/5816 ) determines whether a subnet
should charge cycles. This flag is set by the subnet rental canister and
read/acted on by the execution environment. This PR addresses the latter
part.

### How
The `CyclesAccountManager` is responsible for calculating fees and
handling cycles consumption. It requires two arguments that are not
known at its creation time (and that may change over time):
`subnet_size` and the new `cost_schedule`. Because cost calculations are
used in several different layers (ingress selector, message routing,
execution), this change to the interface of the `CyclesAccountManager`
unfortunately affects a lot of code.
In particular, the registry flag is read every round and passed into the
replicated state every round, because some use sites only have access to
that. Despite this happening every round, this PR does **not** (yet)
address the situation of toggling the flag while the subnet is in
operation.

### Reviewers: 
- The **critical changes** of this PR are in the cycles account manager,
where the cycles fees are calculated.
- The bulk of this change is propagating the information to the use
sites (boilerplate)
- The path of information flow starts at the batch processor, where the
registry flag is read. It is passed to the replicated state and to the
canister manager, where there are some use sites. From the replicated
state, the information is passed to the sandbox safe system state, from
where it is read by the system API. The ingress selector also reads from
the replicated state.

### Tests
- Adapted and extended the cycles account manager tests.
- Added a system test