chore: [EXC-2089] Add logs in system API to track whether bitcoin methods are called on the management canister (#6088)
Calling the bitcoin API via the management canister is
[deprecated](https://internetcomputer.org/docs/references/ic-interface-spec#ic-bitcoin-api)
since a while ago and developers are expected to interact with the
relevant bitcoin canister (mainnet or testnet) directly.

This PR adds a log that would catch any cases where this is still
happening in preparation of fully removing this code from the replica.
Adding a log was preferred over adding some metric as we don't have
access to a `MetricsRegistry` in the system API and propagating that
would be a bunch more work. Instead, adding a log is fairly easier/less
invasive (some info is still propagated to help diagnose but it's quite
limited in scope).